+++ app/src/main/java/uz/coder/muslimcalendar/viewModel/HomeViewModel.kt
@@ -1,9 +1,11 @@
 package uz.coder.muslimcalendar.viewModel
 
+import android.annotation.SuppressLint
 import android.app.Application
 import android.util.Log
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
+import com.google.android.gms.location.LocationServices
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.flow
@@ -19,26 +21,44 @@ class HomeViewModel(private val application: Application):AndroidViewModel(appli
     private val repo = CalendarRepositoryImpl(application)
     private val _state = MutableStateFlow<HomeState>(HomeState.Init)
     val state = _state.asStateFlow()
-    init {
-        remove()
-        loading()
+    private var latitude = 0.0
+    private var longitude = 0.0
+    private val fusedLocationClient by lazy {
+        LocationServices.getFusedLocationProviderClient(application)
     }
-    private fun remove(){
-        viewModelScope.launch {
-            repo.remove()
+    init {
+        initLocation()
+        remove()
+        loadInfo()
+    }
+
+    fun loadInfo() {
+        if (application.isConnected()){
+            loadInformationFromInternet(latitude, longitude)
         }
     }
-    fun loading(){
-        try {
-            loadInformationFromInternet()
-        }catch (_:Exception){}
+
+    @SuppressLint("MissingPermission")
+    private fun initLocation() {
+        fusedLocationClient.lastLocation.addOnSuccessListener {
+            latitude = it.latitude
+            longitude = it.longitude
+        }
     }
-    private fun loadInformationFromInternet() {
+
+    private fun remove(){
+        if (application.isConnected()){
+            viewModelScope.launch {
+                repo.remove()
+            }
+        }
+    }
+    fun loadInformationFromInternet(latitude: Double, longitude: Double) {
         viewModelScope.launch {
             Log.d("TAG", "loadInformationFromInternet: ")
             if (application.isConnected()){
                 _state.value = HomeState.Loading
-                repo.loading()
+                repo.loading(latitude, longitude)
             }else{
                 _state.value = HomeState.Error(application.getString(R.string.no_internet))
             }
@@ -50,9 +70,9 @@ class HomeViewModel(private val application: Application):AndroidViewModel(appli
             repo.presentDay().collect{
                 _state.value = HomeState.Success(listOf(
                     Item(application.getString(R.string.bomdod), it.tongSaharlik),
-                    Item(application.getString(R.string.quyosh), it.quyosh),
                     Item(application.getString(R.string.peshin), it.peshin),
                     Item(application.getString(R.string.asr), it.asr),
+                    Item(application.getString(R.string.quyoshBotishi), it.sunSet),
                     Item(application.getString(R.string.shom), it.shomIftor),
                     Item(application.getString(R.string.xufton), it.hufton)))
             }
@@ -61,7 +81,7 @@ class HomeViewModel(private val application: Application):AndroidViewModel(appli
 
     fun day() = flow {
         repo.presentDay().collect{
-            emit(Date(it.day, it.month-1, it.weekday, it.hijriDay, it.hijriMonth, it.region))
+            emit(Date(it.day, it.month-1, it.weekday, it.hijriDay, it.hijriMonth))
         }
     }
